+++ src/main/native/smithwaterman/PairWiseSW.h
@@ -428,7 +428,7 @@ void inline getCIGAR(SeqPair *p, int16_t *cigarBuf_, int32_t cigarBufLength, int
 }
 
 
-int32_t CONCAT(runSWOnePairBT_,SIMD_ENGINE)(int32_t match, int32_t mismatch, int32_t open, int32_t extend,uint8_t *seq1, uint8_t *seq2, int32_t len1, int32_t len2, int8_t overhangStrategy, char *cigarArray, int32_t cigarLen, int16_t *cigarCount, int32_t *offset)
+int32_t CONCAT(runSWOnePairBT_,SIMD_ENGINE)(int32_t match, int32_t mismatch, int32_t open, int32_t extend,uint8_t *seq1, uint8_t *seq2, int16_t len1, int16_t len2, int8_t overhangStrategy, char *cigarArray, int32_t cigarLen, int16_t *cigarCount, int32_t *offset)
 {
     int32_t  w_match = match;
     int32_t  w_mismatch = mismatch;
@@ -459,8 +459,8 @@ int32_t CONCAT(runSWOnePairBT_,SIMD_ENGINE)(int32_t match, int32_t mismatch, int
     p.seq1 = seq1;
     p.seq2 = seq2;
     // len1 and len2 are bounded by the length of the input with is at most 16 bit.
-    p.len1 = (int16_t)len1;
-    p.len2 = (int16_t)len2;
+    p.len1 = len1;
+    p.len2 = len2;
     p.overhangStrategy = overhangStrategy;
     p.btrack = backTrack_;
     p.cigar = cigarArray;
